updating the datatypes in batchPayment

+++ backend/payment_processing/src/main/java/com/devsquad/payment_processing/model/BatchPaymentRecipient.java
@@ -3,6 +3,8 @@ package com.devsquad.payment_processing.model;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public class BatchPaymentRecipient {
 
     @NotNull(message = "Receiver account number is required")
@@ -10,13 +12,13 @@ public class BatchPaymentRecipient {
 
     @NotNull(message = "Amount is required")
     @Positive(message = "Amount must be positive")
-    private Double amount;
+    private BigDecimal amount;
 
     private String description;  // Optional per-recipient description
 
     public BatchPaymentRecipient() {}
 
-    public BatchPaymentRecipient(Long receiverAccountNumber, Double amount, String description) {
+    public BatchPaymentRecipient(Long receiverAccountNumber, BigDecimal amount, String description) {
         this.receiverAccountNumber = receiverAccountNumber;
         this.amount = amount;
         this.description = description;
@@ -25,8 +27,8 @@ public class BatchPaymentRecipient {
     public Long getReceiverAccountNumber() { return receiverAccountNumber; }
     public void setReceiverAccountNumber(Long receiverAccountNumber) { this.receiverAccountNumber = receiverAccountNumber; }
 
-    public Double getAmount() { return amount; }
-    public void setAmount(Double amount) { this.amount = amount; }
+    public BigDecimal getAmount() { return amount; }
+    public void setAmount(BigDecimal amount) { this.amount = amount; }
 
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
